+++ src/options/components/single-processor-tab.tsx
@@ -254,6 +254,12 @@ export function SingleProcessorTab() {
                 progressiveInterlaced: formatOptions.png.progressiveInterlaced,
                 oxipngCompression: formatOptions.png.oxipngCompression
               }
+            : undefined,
+        tiff:
+          targetFormat === "tiff"
+            ? {
+                colorMode: formatOptions.tiff.colorMode
+              }
             : undefined
       },
       resize: { mode: "none" }